backend/requirements.txt
Flask==3.0.2
Flask_Cors==4.0.0
Flask_Mail==0.9.1
msal==1.28.0
openpyxl==3.1.2
pandas==2.2.1
python-dotenv==1.0.1
Requests==2.31.0
